moved the component to emojis folder
